+++ src/container/class-container.php
@@ -1,4 +1,4 @@
-<?php namespace Fulcrum\Container;
+<?php 
 
 /**
  * Dependency Injection Container - extends the functionality of Pimple DI Container.
@@ -9,7 +9,9 @@
  * @link        http://hellofromtonya.github.io/Fulcrum/
  * @license     GPL-2.0+
  */
+namespace Fulcrum\Container;
 
+use Fulcrum\Support\Exceptions\Configuration_Exception;
 use Pimple\Container as Pimple;
 
 class Container extends Pimple implements Container_Contract {
@@ -30,11 +32,14 @@ class Container extends Pimple implements Container_Contract {
 	 *
 	 * @since 1.0.0
 	 *
-	 * @return self
-	 */
-	public function __construct() {
+	 * @param array $values Array of values to load into the container upon instantiation
+	 * 
+	 * @return Container
+	 */ 
+	public function __construct( array $values = array() ) {
 		self::$instance = $this;
-		parent::__construct( $this->config->initial_parameters );
+		
+		parent::__construct( $values );
 	}
 
 	/****************************
@@ -88,12 +93,28 @@ class Container extends Pimple implements Container_Contract {
 	 * @return mixed
 	 */
 	public function register_concrete( array $config, $unique_id ) {
-		$this[ $unique_id ] = $config['concrete'];
-
-		if ( ! $config['autoload'] ) {
-			return;
+		if ( ! $this->is_concrete_config_valid( $config, $unique_id ) ) {
+			return false;
 		}
 
+		$this[ $unique_id ] = $config['concrete'];
+
+		if ( $config['autoload'] ) {
+			return $this->autoload_the_concrete( $config, $unique_id );
+		}
+	}
+
+	/**
+	 * Autoload the concrete into the container.
+	 *
+	 * @since 1.1.1
+	 *
+	 * @param array $config
+	 * @param string $unique_id
+	 *
+	 * @return mixed
+	 */
+	protected function autoload_the_concrete( array $config, $unique_id ) {
 		if ( true === $config['autoload'] ) {
 			return $this[ $unique_id ];
 		}
@@ -101,5 +122,43 @@ class Container extends Pimple implements Container_Contract {
 		if ( is_callable( $config['autoload'] ) ) {
 			call_user_func( $config['autoload'], $this[ $unique_id ] );
 		}
+	}
+
+	/**
+	 * Checks if the concrete's config is valid.
+	 *
+	 * @since 1.0.0
+	 *
+	 * @param array $config
+	 *
+	 * @return bool
+	 *
+	 * @throws Configuration_Exception
+	 */
+	protected function is_concrete_config_valid( array $config, $unique_id ) {
+		$default = array(
+			'autoload' => false,
+			'concrete' => '',
+		);
+
+		$differences = array_diff_key( $default, $config );
+
+		if ( ! empty( $differences ) ) {
+			throw new Configuration_Exception(
+				sprintf( __( 'The configuration provided for the unique ID of [ %s ] is not valid.', 'fulcrum' ),
+					$unique_id
+				)
+			);
+		}
+
+		if ( ! array_key_exists( 'concrete', $config ) || ! is_callable( $config['concrete'] ) ) {
+			throw new Configuration_Exception(
+				sprintf( __( 'The concrete for the unique ID of [ %s ] is not callable.', 'fulcrum' ),
+					$unique_id
+				)
+			);
+		}
+
+		return true;
 	}
 }