+++ src/custom/post-type/config/defaults.php
@@ -4,7 +4,7 @@
  * Custom post type defaults.
  *
  * @package     Fulcrum\Custom\Post_Type\Config
- * @since       1.0.0
+ * @since       1.1.1
  * @author      hellofromTonya
  * @link        http://hellofromtonya.github.io/Fulcrum/
  * @license     GPL-2.0+
@@ -13,7 +13,9 @@
 namespace Fulcrum\Custom\Post_Type\Config;
 
 return array(
+	'post_type_name' => '',
 	'args'           => array(),
 	'columns_filter' => array(),
 	'columns_data'   => array(),
+	'add_feed'       => false,
 );