+++ src/custom/shortcode/class-shortcode.php
@@ -12,8 +12,6 @@
 
 namespace Fulcrum\Custom\Shortcode;
 
-use InvalidArgumentException;
-use RuntimeException;
 use Fulcrum\Config\Config_Contract;
 
 class Shortcode implements Shortcode_Contract {
@@ -24,6 +22,13 @@ class Shortcode implements Shortcode_Contract {
 	 * @var Config_Contract
 	 */
 	protected $config;
+
+	/**
+	 * No view is required flag
+	 *
+	 * @var bool
+	 */
+	protected $no_view_is_required = false;
 
 	/**
 	 * Shortcode attributes
@@ -45,13 +50,16 @@ class Shortcode implements Shortcode_Contract {
 	 * @since 1.0.0
 	 *
 	 * @param Config_Contract $config Runtime configuration parameters.
+	 * @param Config_Validator Validator
 	 */
-	public function __construct( Config_Contract $config ) {
+	public function __construct( Config_Contract $config, Config_Validator $validator ) {
+		if ( ! $validator->is_config_valid( $config ) ) {
+			return;
+		}
+
 		$this->config = $config;
 
-		if ( $this->is_config_valid() ) {
-			add_shortcode( $this->config->shortcode, array( $this, 'render_callback' ) );
-		}
+		add_shortcode( $this->config->shortcode, array( $this, 'render_callback' ) );
 	}
 
 	/**
@@ -96,9 +104,9 @@ class Shortcode implements Shortcode_Contract {
 
 	/**
 	 * Get the content.  This method processes the content by passing it
-	 * through the shortcode function and escaping through `wp_kses_post`.
+	 * through the shortcode function and escaping through a sanitizing filter.
 	 *
-	 * @since 1.0.0
+	 * @since 1.1.1
 	 *
 	 * @return string
 	 */
@@ -109,7 +117,7 @@ class Shortcode implements Shortcode_Contract {
 
 		$content = do_shortcode( $this->content );
 
-		$filter  = $this->config->has( 'content_filter' ) ? $this->config->content_filter : 'wp_kses_post';
+		$filter  = $this->config->content_filter;
 		$content = $filter( $content );
 
 		return $content;
@@ -143,40 +151,5 @@ class Shortcode implements Shortcode_Contract {
 		}
 
 		return ' ' . esc_attr( $this->atts['class'] );
-	}
-
-	/**
-	 * Checks if the config is valid to start
-	 *
-	 * @since 1.0.0
-	 *
-	 * @return bool
-	 * @throws InvalidArgumentException
-	 * @throws RuntimeException
-	 */
-	protected function is_config_valid() {
-		if ( ! $this->config->has( 'shortcode' ) ||
-		     ! $this->config->is_array( 'defaults' ) ||
-		     ! $this->config->has( 'view' )
-		) {
-			throw new InvalidArgumentException( __( 'Invalid config for shortcode.', 'fulcrum' ) );
-		}
-
-		if ( ! $this->is_no_view_required() && ! is_readable( $this->config->view ) ) {
-			throw new RuntimeException( sprintf( __( 'The specified view file [%s] is not readable.', 'fulcrum' ), $this->config->view ) );
-		}
-
-		return true;
-	}
-
-	/**
-	 * Checks if a no view is required.
-	 *
-	 * @since 1.0.1
-	 *
-	 * @return bool
-	 */
-	protected function is_no_view_required() {
-		return $this->config->has( 'no_view' ) && $this->config->no_view;
 	}
 }