+++ src/custom/shortcode/config/defaults.php
@@ -1,0 +1,21 @@
+<?php
+
+/**
+ * Shortcode defaults.
+ *
+ * @package     Fulcrum\Custom\Shortcode\Config
+ * @since       1.1.1
+ * @author      hellofromTonya
+ * @link        http://hellofromtonya.github.io/Fulcrum/
+ * @license     GPL-2.0+
+ */
+
+namespace Fulcrum\Custom\Shortcode\Config;
+
+return array(
+	'shortcode'      => '',
+	'no_view'        => false,
+	'view'           => '',
+	'defaults'       => array(),
+	'content_filter' => 'wp_kses_post',
+);